% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn_to_cpprouting.R
\name{sfn_to_cpprouting}
\alias{sfn_to_cpprouting}
\title{Convert an sfnetwork object to a cppRouting graph}
\usage{
sfn_to_cpprouting(sfn, directed = FALSE)
}
\arguments{
\item{sfn}{An \code{sfnetwork} object.}

\item{directed}{Logical, whether the graph should be directed. Defaults to \code{FALSE}.}
}
\value{
A \code{cppRouting} graph object.
}
\description{
This function takes an \code{sfnetwork} object and converts it into a graph
format that is usable by the \code{cppRouting} package.
}
\examples{
\dontrun{
if (requireNamespace("sfnetworks", quietly = TRUE) && requireNamespace("sf", quietly = TRUE)) {
  data(sr_data_osm)
  sfn = sfnetworks::as_sfnetwork(sr_data_osm)
  graph = sfn_to_cpprouting(sfn)
}
}
}
