% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_data_osm.R
\docType{data}
\name{sr_data_osm}
\alias{sr_data_osm}
\title{Sample road network data for Leeds from OpenStreetMap}
\format{
An \code{sf} data frame with road segments:
\describe{
\item{geometry}{\code{LINESTRING} geometry of the road segment.}
\item{highway}{The type of highway (e.g., 'residential', 'primary').}
\item{...}{Other columns from OpenStreetMap with less than 90\% missing values.}
}
}
\source{
\url{https://www.openstreetmap.org/}
}
\usage{
sr_data_osm
}
\description{
A dataset containing road network data for Leeds, UK, extracted from
OpenStreetMap using the \code{osmactive} package. The data has been
processed for routing applications. It has been filtered to remove columns
with a high proportion of missing values (more than 90\%) and the geometry
has been cast to \code{LINESTRING}.
}
\keyword{datasets}
