% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_route.R
\name{sr_route}
\alias{sr_route}
\title{Get a route from an sfnetwork}
\usage{
sr_route(sfn, from, to)
}
\arguments{
\item{sfn}{An \code{sfnetwork} object.}

\item{from}{The node ID of the origin.}

\item{to}{The node ID of the destination.}
}
\value{
An \code{sf} object representing the route edges.
}
\description{
This function calculates a route between an origin and a destination node
in an \code{sfnetwork} object.
}
\examples{
\dontrun{
if (requireNamespace("sfnetworks", quietly = TRUE) && requireNamespace("sf", quietly = TRUE)) {
  data(sr_data_osm)
  sfn = sfnetworks::as_sfnetwork(sr_data_osm)
  # Ensure nodes have IDs
  sfn = sfn |>
    tidygraph::activate("nodes") |>
    dplyr::mutate(ID = dplyr::row_number())
  from_node = 1
  to_node = 100
  route_sf = sr_route(sfn, from = from_node, to = to_node)
  plot(route_sf)
}
}
}
